fix warnings
git-svn-id: http://svn.code.sf.net/p/gpac/code/trunk/gpac@5391 63c20433-aa62-49bd-875c-5a186b69a8fb

+++ src/media_tools/media_export.c
@@ -1187,7 +1187,7 @@ GF_Err gf_media_export_native(GF_MediaExporter *dumper)
 		qcp_type = 0;
 	} else if (m_stype==GF_ISOM_SUBTYPE_STXT) {
 		if (stxtcfg) {
-			gf_bs_write_data(bs, stxtcfg, strlen(stxtcfg));
+			gf_bs_write_data(bs, stxtcfg, (u32)strlen(stxtcfg));
 		}
 	}
